Date/time data types as custom types, and refactor typeclass handling to generate code which compiles faster
Finally solve the problem of creating working db codecs from temporal types by going through string encoding.
It's not possible to correctly roundtrip these data types through the PG driver and jdbc layers.

Note that it could be done with orphan typeclasses except for the DSL, which is why I needed to go through wrapper types in the end :(

for circe:
- use `Encoder.forProductN` and same for `Decoder` when possible (n <= 22)
- use exception flow for n > 22 to save on flatMaps

for play:
- use apply methods instead of SAM
- avoid some implicit machinery by using `Json.toJson` with explicit types

Details:
- core: fix handling of `TimeTz`
- unify instance names across all libs
- revamp typeclasses
- introduce instances as nodes in tree
- split multiple instances of multiple type classes into multiple definitions
- always sort instances by name
- reorg DbLib files
- never use `Format` for playjson
- allow empty package object