towards custom types
I thought it was better to communicate this in code than to try to explain it.

the custom types rely on the default types produced and consumed by the postgres driver, so we effectively just use `{get|set}Object`.
Then there are functions which transform those types to/from the custom types exposed in typo.

I'm not entirely sure this will all work as I couldn't test it, but we can iterate on it

note that for unary types we can produce `JdbcEncoder` from a `Setter`:
```scala
JdbcEncoder.singleParamEncoder(jdbcSetter)
```