Add support for generating code into a build with multiple projects
It's very common that you don't want to expose all the tables in one place in your system.
You also don't want duplication of generated code or more than one script to generate database code.

The solution is to pass a `ProjectGraph` structure to `generateFromDb` instead,
in which you encode the structure of the (relevant) projects in your build.

Dependencies between projects are encoded by nesting in the tree you pass (each node has a `downstream` member),
with the root being the uppermost one.

If multiple downstream projects want to generate the same code, it'll be pulled up to the level necessary to become visible for all of them.