Revamp SQL generation
The previous method for SQL generation was not powerful enough

## Create scopes for all levels

The previous method failed to create "scopes" for all joins, which meant that predicates, sorting, limits and so on wouldn't apply where usage of the DSL indicated they would.

Now we create these scopes with CTEs at all levels of the query.
CTEs should be fine, performance-wise, as they were fixed for PG12

## Better compositionality
The previous method failed when performing (joined query).join(joined query), it only worked with relations on the left side. This is also fixed

## Refactoring
The old SQL generation code was neither understandable nor correct. Now I hope it's both

## Lost predicates for intermediate joins
Since the old code was hard to understand,